// frontend/src/entity/databases/model/Database.ts
import type { Notifier } from '../../notifiers';
import type { DatabaseType } from './DatabaseType';
import type { HealthStatus } from './HealthStatus';
import type { MariadbDatabase } from './mariadb/MariadbDatabase';
import type { MongodbDatabase } from './mongodb/MongodbDatabase';
import type { MysqlDatabase } from './mysql/MysqlDatabase';
import type { MysqlPhysicalDatabase } from './mysql/MysqlPhysicalDatabase';
import type { PostgresqlLogicalDatabase } from './postgresql/PostgresqlLogicalDatabase';
import type { PostgresqlPhysicalDatabase } from './postgresql/physical/PostgresqlPhysicalDatabase';

export interface Database {
  id: string;
  name: string;
  workspaceId: string;
  type: DatabaseType;

  postgresqlLogical?: PostgresqlLogicalDatabase;
  postgresqlPhysical?: PostgresqlPhysicalDatabase;
  mysql?: MysqlDatabase;
  mysqlPhysical?: MysqlPhysicalDatabase;
  mariadb?: MariadbDatabase;
  mongodb?: MongodbDatabase;

  notifiers: Notifier[];

  lastBackupTime?: Date;
  lastBackupErrorMessage?: string;

  healthStatus?: HealthStatus;
}